Разработка экрана "Общие параметры" в адмике (продолжение).

+++ app/src/main/java/com/mu/tote2026/presentation/screen/game/GameViewModel.kt
@@ -310,9 +310,9 @@ class GameViewModel @Inject constructor(
         }
 
         val coefficient = if (game.groupId.toInt() <= GROUPS_COUNT) {
-            if (pointsSum > 0) common.groupPrizeFund / pointsSum else 0.0
+            if (pointsSum > 0) common.groupPrizeFund.toDouble() / pointsSum else 0.0
         } else {
-            if (pointsSum > 0) common.playoffPrizeFund / pointsSum else 0.0
+            if (pointsSum > 0) common.playoffPrizeFund.toDouble() / pointsSum else 0.0
         }
 
         game.stakes.forEachIndexed { idx, stake ->
